+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- 将部分运维指令中的通用方法提升到 `CommandUtil` 中。
+  - com.dwarfeng.fdr.impl.service.telqos.CommandUtil.cropData。
+
 - 增加预设的运维指令。
   - com.dwarfeng.springtelqos.api.integration.system.UptimeCommand。
   - com.dwarfeng.springtelqos.api.integration.system.JmxRemoteCommand。